getUsersFollowers now matches the database schema

+++ sql/generatePostgres.sql
@@ -1,12 +1,12 @@
 CREATE TABLE grp(
-	groupId BIGINT PRIMARY KEY,
+	groupId SERIAL PRIMARY KEY,
 	name VARCHAR(50)
 );
 CREATE TABLE usr(
 	userId BIGINT PRIMARY KEY
 );
 CREATE TABLE userInGroup(
-	groupId BIGINT REFERENCES grp (groupId), 
+	groupId INT REFERENCES grp (groupId), 
 	userId 	BIGINT REFERENCES usr (userId),
 	PRIMARY KEY(groupId,userId)
 );
